Move tests out of distributed package and flatten project
In the spirit of moving tests outside of the distributed Python package, move
the tests.py module to the top level of the project. The unnecessary
intermediate src directory has been removed in favor of a flatter project. All
references to files and tests have been updated with this new structure.